More refactoring
Rust-only code is split off into a separate file, and lib.rs only
contains the Python shim layer.
